Swap MockStdio to FakeStdio in tests
