+++ packages/flutter_tools/test/general.shard/analytics_test.dart
@@ -24,7 +24,7 @@ import 'package:usage/usage_io.dart';
 
 import '../src/common.dart';
 import '../src/context.dart';
-import '../src/mocks.dart';
+import '../src/fakes.dart';
 
 void main() {
   setUpAll(() {
@@ -150,7 +150,7 @@ void main() {
 
   group('analytics with mocks', () {
     MemoryFileSystem memoryFileSystem;
-    MockStdio mockStdio;
+    FakeStdio fakeStdio;
     Usage mockUsage;
     SystemClock mockClock;
     Doctor mockDoctor;
@@ -158,7 +158,7 @@ void main() {
 
     setUp(() {
       memoryFileSystem = MemoryFileSystem.test();
-      mockStdio = MockStdio();
+      fakeStdio = FakeStdio();
       mockUsage = MockUsage();
       mockClock = MockClock();
       mockDoctor = MockDoctor();
@@ -264,7 +264,7 @@ void main() {
           'FLUTTER_ANALYTICS_LOG_FILE': 'analytics.log',
         },
       ),
-      Stdio: () => mockStdio,
+      Stdio: () => fakeStdio,
     });
 
     testUsingContext('event sends localtime', () async {
@@ -294,7 +294,7 @@ void main() {
           'FLUTTER_ANALYTICS_LOG_FILE': 'analytics.log',
         },
       ),
-      Stdio: () => mockStdio,
+      Stdio: () => fakeStdio,
     });
   });
