+++ packages/flutter_tools/test/general.shard/android/android_workflow_test.dart
@@ -16,7 +16,8 @@ import 'package:mockito/mockito.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
-import '../../src/mocks.dart' show MockAndroidSdk, MockProcess, MockProcessManager, MockStdio;
+import '../../src/fakes.dart';
+import '../../src/mocks.dart' show MockAndroidSdk, MockProcess, MockProcessManager;
 import '../../src/testbed.dart';
 
 class MockAndroidSdkVersion extends Mock implements AndroidSdkVersion {}
@@ -27,7 +28,7 @@ void main() {
   Logger logger;
   MemoryFileSystem fileSystem;
   MockProcessManager processManager;
-  MockStdio stdio;
+  FakeStdio stdio;
 
   setUp(() {
     sdk = MockAndroidSdk();
@@ -35,7 +36,7 @@ void main() {
     fileSystem.directory('/home/me').createSync(recursive: true);
     logger = BufferLogger.test();
     processManager = MockProcessManager();
-    stdio = MockStdio();
+    stdio = FakeStdio();
   });
 
   MockProcess Function(List<String>) processMetaFactory(List<String> stdout) {
