+++ packages/flutter_tools/test/src/fakes.dart
@@ -5,6 +5,7 @@
 import 'dart:async';
 
 import 'package:flutter_tools/src/base/file_system.dart';
+import 'package:flutter_tools/src/base/io.dart';
 import 'package:flutter_tools/src/base/logger.dart';
 import 'package:flutter_tools/src/base/os.dart';
 import 'package:flutter_tools/src/cache.dart';
@@ -62,4 +63,21 @@ class FakeDyldEnvironmentArtifact extends ArtifactSet {
   @override
   Future<void> update(ArtifactUpdater artifactUpdater, Logger logger, FileSystem fileSystem, OperatingSystemUtils operatingSystemUtils) async {
   }
+}
+
+class FakeStdio extends Stdio {
+  FakeStdio({this.hasFakeTerminal});
+
+  final bool hasFakeTerminal;
+
+  @override
+  bool get hasTerminal => hasFakeTerminal;
+
+  @override
+  int get terminalColumns => hasFakeTerminal ? 80 : null;
+
+  @override
+  int get terminalLines => hasFakeTerminal ? 24 : null;
+  @override
+  bool get supportsAnsiEscapes => hasFakeTerminal;
 }
